pyproject.toml
[tool.poetry]
name = "tortoise-serializer"
version = "1.3.1"
description = "Pydantic serialization for tortoise-orm"
authors = ["Sebastien Nicolet <snicolet95@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.10.3"
frozendict = "^2.4.6"
structlog = "^24.4.0"
tortoise-orm = "^0.23.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.30.0"
pytest = "^8.3.4"
anyio = "^4.7.0"
pytest-asyncio = "^0.25.0"
ruff = "^0.8.4"
pytest-cov = "^6.0.0"
coverage = "^7.6.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
target-version = "py312"
line-length = 79
indent-width = 4
